rename ui and svg to views
