+++ app/view-html-ui.js
@@ -1,5 +1,5 @@
 //
-//	html-ui
+//	view html-ui
 //
 
 import { HTMLApp } from "../[library]/HTMLApp.js";
